move around png writing functions and setup for png drawing
