+++ src/subcommand/viz_main.cpp
@@ -9,60 +9,11 @@
 #include <limits>
 #include "picosha2.h"
 #include <iostream>
+#include "algorithms/draw.hpp"
 
 //#define debug_odgi_viz
 
 namespace odgi {
-
-    namespace png {
-
-        /*
-        3 ways to encode a PNG from RGBA pixel data to a file (and 2 in-memory ways).
-        NOTE: this samples overwrite the file or test.png without warning!
-        */
-
-        //g++ lodepng.cpp examples/example_encode.cpp -I./ -ansi -pedantic -Wall -Wextra -O3
-
-        //Example 1
-        //Encode from raw pixels to disk with a single function call
-        //The image argument has width * height RGBA pixels or width * height * 4 bytes
-        void encodeOneStep(const char *filename, std::vector<unsigned char> &image, unsigned width, unsigned height) {
-            //Encode the image
-            unsigned error = lodepng::encode(filename, image, width, height);
-
-            //if there's an error, display it
-            if (error) std::cout << "encoder error " << error << ": " << lodepng_error_text(error) << std::endl;
-        }
-
-        //Example 2
-        //Encode from raw pixels to an in-memory PNG file first, then write it to disk
-        //The image argument has width * height RGBA pixels or width * height * 4 bytes
-        void encodeTwoSteps(const char *filename, std::vector<unsigned char> &image, unsigned width, unsigned height) {
-            std::vector<unsigned char> png;
-
-            unsigned error = lodepng::encode(png, image, width, height);
-            if (!error) lodepng::save_file(png, filename);
-
-            //if there's an error, display it
-            if (error) std::cout << "encoder error " << error << ": " << lodepng_error_text(error) << std::endl;
-        }
-
-        //Example 3
-        //Save a PNG file to disk using a State, normally needed for more advanced usage.
-        //The image argument has width * height RGBA pixels or width * height * 4 bytes
-        void encodeWithState(const char *filename, std::vector<unsigned char> &image, unsigned width, unsigned height) {
-            std::vector<unsigned char> png;
-            lodepng::State state; //optionally customize this one
-
-            unsigned error = lodepng::encode(png, image, width, height, state);
-            if (!error) lodepng::save_file(png, filename);
-
-            //if there's an error, display it
-            if (error) std::cout << "encoder error " << error << ": " << lodepng_error_text(error) << std::endl;
-        }
-
-    }
-
 
     using namespace odgi::subcommand;
